Move Signup* types to their own package
Needed in order to import them from several modules.
